requirements.txt
Mastodon.py==1.6.3
peewee==3.15.4
